feat: add controller and services for registration
The controller and services layer for user registration.
The controller uses the arguments of the request body to be used as the
data of the new user. The arguments of the request body are validated
by middleware meant to validate the arguments for required for
registration.

The services for defines methods for the controller layer doing
exactly as their name implies:
isValidEmail, isValidPassword, findUser, generateFriendlyId and insertUser.